Ignore updating counter if it drops in value.
This can happen if dnsmasq container that sidecar is
pulling metrics from, restarts, but sidecar does not restart.
One way this can happen is if liveness probes fail for dnsmasq and
it gets restarted.
Sidecar will panic since the counter metrics are reset for the
newly started container. This change skips the update until the
metrics are caught up. This does introduce an inconsistency in the metrics
for a brief interval.