Fix skipteardown flag not completely avoiding teardown
When using an HA setup of nodecache, it is imperative to avoid that one
instance of nodecache tear down networking at any time, to avoid disrupting
functionalities for the second instance, which is sharing the same interface
and iptables rules.

This fix addresses the issue of nodecache trying to tear down networking
at startup, right before ensuring it is setup again. When `--skipteardown`
flag is passed, nodecache should not try to do so, but instead simply
ensure that networking is correctly setup.